==> Scripts/1b-gene-expression-PCA-Pacu.Rmd <==
---
title: "edgeR Workflow for gene expression PCA of *P. acuta*"
author: "Erin Chille"
date: "02/16/2023"
output:
  pdf_document: default
  html_notebook: default
---


```{r setup, include=FALSE}
rm(list = ls()) #clear environment
```

*Note: Open Rproj first, then script. To easily use relative paths, click the down button next to knit and then click "Knit Directory --> Project Directory". This should make loading and saving files much easier.*

## 1. Load packages and input data

Load packages
```{r, message=FALSE, warning=FALSE}
library(edgeR, quietly = TRUE) #edgeR-v3.30.3
library(vegan, quietly = TRUE)
library(Dune, quietly = TRUE)
library(ggplot2, quietly = TRUE) #ggplot2-v3.3.5
library(tidyverse, quietly = TRUE) #tidyverse-v1.3.1
library(ComplexHeatmap, quietly = TRUE)
```

Load the input file containing the treatment information
```{r}
#treatment information
treatmentinfo <- read.csv("Sample_Info/samples_Pacuta.annotations.txt", header = TRUE, sep = "\t", fileEncoding="UTF-8-BOM") #read in file
top3groups <- c("Group2", "Group3", "Group6")
treatmentinfo <- filter(treatmentinfo, group%in%top3groups)
table(treatmentinfo$group)
```

Load the input file containing the gene count matrix
```{r}
#gene count matrix
gcount <- as.data.frame(read_delim("Genome_Info/Pocillopora_acuta_KBHIv2.gentrome.fa.gz.salmon.numreads.matrix", delim = "\t", col_names = TRUE, show_col_types = FALSE), fileEncoding="UTF-8-BOM") #read in file
rownames(gcount) <- gcount$Name #makes "Name" the rowname
gcount <- gcount[,-c(1)] #drops the "Name" column
gcount <- round(gcount) #round 
dim(gcount); head(gcount)[,1:3] #view dataset attributes
gcount <- gcount[,treatmentinfo$sample]
```

Determine library size
```{r}
libSize.df <- data.frame(libSize=colSums(gcount))
```

Make DGE object
```{r}
DGEdat <- DGEList(counts=as.matrix(gcount), samples=treatmentinfo,
                  group=treatmentinfo$temp)
dim(DGEdat$counts)
```

## 2. Pre-filtering
```{r}
lib.sizes.cpm <- colSums(gcount)/1000000
lib.sizes.cpm <- data_frame(sample=colnames(gcount), lib.size=lib.sizes.cpm)

ggplot(lib.sizes.cpm, aes(y = lib.size, x=sample)) +
  ylab("Library size (per million)") + xlab("Sample") + # Set x and y axis labels
  ylim(3,10)+
  geom_jitter(colour = "cadetblue", alpha = 0.5) +
  geom_boxplot(colour="cadetblue", alpha=0) + 
  theme_bw() + #Set background color
  theme(panel.border = element_blank(), # Set border
                     panel.grid.major = element_blank(), #Set major gridlines
                     panel.grid.minor = element_blank(), #Set minor gridlines
                     axis.line = element_line(colour = "black"), #Set axes color
                      axis.title.x = element_blank(), # No axis title
                      axis.text.x = element_text(color = "White"), #White-out the x-axis label
                      axis.ticks.x = element_blank(), # No ticks on x-axis
                      axis.title.y = element_text(size = 16), #Set Yaxis label size
                     axis.text=element_text(size=12), #Set Y tick label size
                      plot.background=element_blank()) + #Set the plot background
  ggtitle("(b)") + #add a main title
  scale_color_manual(values=c(AMB="cadetblue", l="indianred3", xl="deeppink4")) + #set treatment colors
  theme(plot.title = element_text(face = 'bold', 
                                  size = 12, 
                                  hjust = 0),
        legend.position = "bottom") #set title attributes
mean(lib.sizes.cpm$lib.size)
median(lib.sizes.cpm$lib.size)
10/min(lib.sizes.cpm$lib.size)

keep <- rowSums(cpm(gcount) > 3.33) >= 2
table(keep)
DGEdat <- DGEdat[keep, , keep.lib.sizes=FALSE]
```

##  3. Data normalization  
```{r}
DGEdat <- calcNormFactors(DGEdat)
DGEdat$samples
```

##  4. Plot global gene expression  

Log transform the counts matrix for the next plots
```{r}
DGEdat.cpm <- DGEdat #make a copy the edgeR dataset
DGEdat.cpm$counts <- cpm(DGEdat.cpm$counts, log=TRUE, prior.count=5) #log transform the copy for the next plots
```

Run a principle coordinates analysis
```{r}
d <- data.frame(Group = DGEdat.cpm$samples$group.1, DGEdat.cpm$samples, name = colnames(DGEdat.cpm$counts)) #make a dataframe containing all plotting info
pca <- prcomp(t(DGEdat.cpm$counts[,d$sample])) #calculate eigengenes
pc.data <- summary(out<-prcomp(t(DGEdat.cpm$counts[,d$sample]))); pc.data
plot(out) #theres a lot of variation in the data
DGEdat_PCcor <- lapply(d[,c(7:10,12,14,16,18)], as.factor)
DGEdat_PCcor <- as.tibble(lapply(DGEdat_PCcor, as.numeric))
rownames(DGEdat_PCcor) <- d$sample
all<-cbind(scores(out), DGEdat_PCcor)
correlations<-round(cor(all, method = "spearman"), digits=2)
plot(correlations)
pc.data
write_csv(as.data.frame(correlations), "Output/1a-Pacu-PC-correlations.csv")

pc_cor<-cbind(scores(out), DGEdat_PCcor)
pc.data

kruskal.test(PC7 ~ treatment, data = pc_cor)
kruskal.test(PC1 ~ group.1, data = pc_cor)

pairwise.wilcox.test(pc_cor$PC1,pc_cor$group.1,
                 p.adjust.method = "BH")
```
Visualize first two principle coordinates
```{r}
percentVar <- pca$sdev^2/sum(pca$sdev^2) #save % variation by PC1 and PC2
d <- data.frame(PC1 = pca$x[, 1], PC7 = pca$x[, 7], d) #make a dataframe containing all plotting info

allgenes_PCA <- ggplot(data = d, aes_string(x = "PC1", y = "PC7")) + 
  geom_point(size = 3, color='black', aes(shape=temp, fill=group.1)) +
  scale_shape_manual(guide="legend", values=c('Amb'=21, 'Hot'=24)) +
  scale_fill_manual(guide="legend", values=c('Group2'="#FFBD5C", 'Group3'="#B873FF", 'Group6'="#4EB900")) +
  xlab(paste0("PC1: ", round(percentVar[1] * 100), "% variance")) +
  ylab(paste0("PC7: ", round(percentVar[7] * 100), "% variance")) +
  coord_fixed() +
  theme_bw() + #Set background color
  theme(panel.border = element_blank(), # Set border
        panel.grid.major = element_blank(), #Set major gridlines
        panel.grid.minor = element_blank(), #Set minor gridlines
        axis.line = element_line(colour = "black", size = 0.6), #Set axes color
        plot.background=element_blank(), #Set the plot background
        axis.title = element_text(size = 14)) + #Axis title sizet
  guides(fill = guide_legend(override.aes = list(shape=21))); allgenes_PCA
ggsave("Output/1a-Pacu-edgeR-allsamples-DEGs-PCA.pdf", allgenes_PCA)

```

Plot as a dendrogram
```{r}
tree <- hclust(dist(as.matrix(t(DGEdat.cpm$counts))))
dend <- as.dendrogram(tree)
#pdf("Output/NEW_tree_SNPs_allsamples-allgenes-.pdf", height = 3, width = 20)
plot(dend)
#dev.off()
```

There don't look to be any outliers. 

##  5. Differential gene expression analysis for Diploids versus Triploids

Creating the design matrix
```{r}
design <- model.matrix(~0+DGEdat$samples$group)
levels(DGEdat$samples$group)
colnames(design) <- levels(DGEdat$samples$group)
head(design); dim(design)
```

Estimate dispersion and fit a glmQL model
```{r}
DGEdat <- estimateDisp(DGEdat, design, robust=TRUE)
fit <- glmQLFit(DGEdat, design, robust=TRUE)
head(fit$coefficients)
plotQLDisp(fit)
```

Set contrasts
```{r}
test <- makeContrasts(Hot - Amb, levels = design)
res <- glmQLFTest(fit, contrast = test)
topTags(res)
T1.de <- decideTestsDGE(res, adjust.method="BH", p.value=0.05, lfc=2)
summary(T1.de)
DEGtable <- res$table[T1.de!=0,]
DEGtable$gene_id  <- rownames(DEGtable)
rownames(DEGtable) <- NULL
DEGtable <- cbind(DEGtable, cluster=ifelse(DEGtable$logFC>0, "Up", "Down"))
# write_csv(DEGtable[,c(6,1:5,7)], "Output/Supplementary_table_ploidy_DEGs.csv", col_names = T)
```


##  6. Explore and visualize results

```{r}
plotMD(res)
```


### Plot a heatmap of differentially-expressed genes

We will now transform them with cpm for plotting
```{r}
gcount.cpm <- cpm(gcount, log=TRUE, prior.count=5)
DEGcounts.cpm <- gcount.cpm[DEGtable$gene_id, ]
dim(DEGcounts.cpm)
```

We will now transform them with cpm for plotting
```{r}
gcount.cpm <- cpm(gcount, log=TRUE, prior.count=5)
DEGcounts.cpm <- gcount.cpm[DEGtable$gene_id, ]
dim(DEGcounts.cpm)
```

Make a heatmap 
```{r}
mat <- as.matrix(DEGcounts.cpm - rowMeans(DEGcounts.cpm))
hmTreatment <- subset(DGEdat$samples, select=c("temp"))
hm_ann_col <- HeatmapAnnotation(df=hmTreatment, col = list(temp = c("Amb" = "#1b9e77", "Hot" = "#d95f02"))) #make dataframe for column naming
dend = cluster_within_group(mat, hmTreatment$temp)

DEGheatmap <-  Heatmap(mat, cluster_columns = dend, column_split = 2,
        name = "Gene expression (cpm)", 
        show_row_names = FALSE, 
        top_annotation = hm_ann_col, 
        show_column_names = FALSE, row_dend_side = "left" ,
        column_dend_height = unit(0.5, "in"),
        row_title_side = "right", row_title_rot = 0, row_dend_reorder = TRUE, 
        row_gap = unit(2.5, "mm"), border = TRUE,
        column_names_gp =  gpar(fontsize = 10))

pdf("Output/1a-Pacu-edgeR-Treatment-DEG-heatmap.pdf", width = 10.5, height = 7)
draw(DEGheatmap)
dev.off()
```


### Principle components plot of DEGs
```{r}
pca <- prcomp(t(DEGcounts.cpm)) #calculate eigengenes
id=
percentVar <- pca$sdev^2/sum(pca$sdev^2) #save % variation by PC1 and PC2
d6 <- data.frame(PC1 = pca$x[, 1], PC2 = pca$x[, 2], temp = as.factor(DGEdat.cpm$samples$temp), Group = as.factor(DGEdat.cpm$samples$temp), #make a dataframe containing all plotting info
                DGEdat.cpm$samples, name = colnames(DGEdat.cpm$counts))
DEG_PCA <- ggplot(data = d6, aes_string(x = "PC1", y = "PC2")) + 
  geom_point(size = 4, aes(colour=temp)) + 
  xlab(paste0("PC1: ", round(percentVar[1] * 100), "% variance")) + 
  ylab(paste0("PC2: ", round(percentVar[2] * 100), "% variance")) + 
  coord_fixed() + 
  theme_bw() + #Set background color
  theme(panel.border = element_blank(), # Set border
        panel.grid.major = element_blank(), #Set major gridlines
        panel.grid.minor = element_blank(), #Set minor gridlines
        axis.line = element_line(colour = "black", size = 0.6), #Set axes color
        plot.background=element_blank(), #Set the plot background
        axis.title = element_text(size = 14), #Axis title size
        axis.text = element_blank());DEG_PCA #Axis text size and view plot
ggsave("Output/1a-Pacu-edgeR-allsamples-DEGs-PCA.pdf", DEG_PCA)
```
